pyproject.toml
[project]
name = "financial-trading-server"
version = "0.1.0"
description = ""
authors = [
    {name = "michael",email = "7471919@naver.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "langgraph (>=0.3.10,<0.4.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
